Move solver_test_utils to test/common for reuse in regression tests.
