+++ test/common/solver_test_utils.hpp
