+++ test/integration/solvers/H3LAPD/test_H3LAPD.hpp
@@ -5,7 +5,7 @@
 
 #include "EquationSystems/DriftReducedSystem.hpp"
 #include "EquationSystems/HW2Din3DSystem.hpp"
-#include "solver_test_utils.hpp"
+#include "../../common/solver_test_utils.hpp"
 #include "solvers/solver_callback_handler.hpp"
 #include "solvers/solver_runner.hpp"
